requirements.txt
numpy
sounddevice
soundfile
groq
elevenlabs
python-dotenv
fastapi
uvicorn
deepface
easyocr
pdf2image
opencv-python
pydantic
requests
python-multipart
httpx
sqlalchemy
asyncpg
alembic
xgboost
joblib
scikit-learn
pandas
scipy
pdfplumber
PyMuPDF
fpdf2
tf-keras
celery